pyproject.toml
[tool.pdm]

[project]
name = "la-readme-stats"
version = "0.1.0"
description = ""
authors = [
    {name = "thiagola92", email = "thiagola92@gmail.com"},
]
dependencies = [
    "PyGithub>=1.57",
]
requires-python = ">=3.11"
license = {text = "MIT"}
